hetzner/requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
fabric==2.6.0
idna==3.2
invoke==1.6.0
paramiko==2.7.2
pathlib2==2.3.6
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
